fix(runtime): don't run the event loop in case of compilation error
