+++ packages/runtime/tests/runtime.rs
@@ -251,12 +251,17 @@ async fn return_uint8array() {
 }
 
 #[tokio::test(flavor = "multi_thread")]
-async fn promise() {
+async fn console_log() {
     setup();
     let mut isolate = Isolate::new(IsolateOptions::new(
-        "export async function handler() {
-    const body = await fetch('http://google.com').then((res) => res.text());
-    return new Response(body);
+        "export function handler() {
+    const types = ['log', 'info', 'debug', 'error', 'warn'];
+
+    types.forEach(type => {
+        console[type]('Hello world!')
+    })
+
+    return new Response('');
 }"
         .into(),
     ));
@@ -264,75 +269,5 @@ async fn promise() {
     assert_eq!(
         isolate.run(Request::default()).await.0,
         RunResult::Response(Response::default())
-    );
-}
-
-#[tokio::test(flavor = "multi_thread")]
-async fn handler_reject() {
-    setup();
-    let mut isolate = Isolate::new(IsolateOptions::new(
-        "export function handler() {
-    throw new Error('Rejected');
-}"
-        .into(),
-    ));
-
-    assert_eq!(
-        isolate.run(Request::default()).await.0,
-        RunResult::Error("Uncaught Error: Rejected".into()),
-    );
-}
-
-#[tokio::test(flavor = "multi_thread")]
-async fn timeout_reached() {
-    setup();
-    let mut isolate = Isolate::new(IsolateOptions::new(
-        "export function handler() {
-    while(true) {}
-    return new Response('Should not be reached');
-}"
-        .into(),
-    ));
-
-    assert_eq!(
-        isolate.run(Request::default()).await.0,
-        RunResult::Timeout(),
-    );
-}
-
-#[tokio::test(flavor = "multi_thread")]
-async fn memory_reached() {
-    setup();
-    let mut isolate = Isolate::new(
-        IsolateOptions::new(
-            "export function handler() {
-    const storage = [];
-    const twoMegabytes = 1024 * 1024 * 2;
-    while (true) {
-        const array = new Uint8Array(twoMegabytes);
-        for (let ii = 0; ii < twoMegabytes; ii += 4096) {
-        array[ii] = 1; // we have to put something in the array to flush to real memory
-        }
-        storage.push(array);
-    }
-    return new Response('Should not be reached');
-}"
-            .into(),
-        )
-        // Increase timeout for CI
-        .with_timeout(1000),
-    );
-
-    assert_eq!(
-        isolate
-            .run(Request {
-                body: "".into(),
-                headers: HashMap::new(),
-                method: Method::GET,
-                url: "".into(),
-            })
-            .await
-            .0,
-        RunResult::MemoryLimit(),
     );
 }
